fix: Make 2I and 2J tests dRep dependent
