+++ tests/govtool-frontend/playwright/tests/2-delegation/delegation.loggedin.spec.ts
@@ -1,9 +1,7 @@
-import { dRep01Wallet, user01Wallet } from "@constants/staticWallets";
+import { user01Wallet } from "@constants/staticWallets";
 import { test } from "@fixtures/walletExtension";
 import { setAllureEpic } from "@helpers/allure";
-import { ShelleyWallet } from "@helpers/crypto";
 import { isMobile } from "@helpers/mobile";
-import extractDRepFromWallet from "@helpers/shellyWallet";
 import DRepDirectoryPage from "@pages/dRepDirectoryPage";
 import { expect } from "@playwright/test";
 
@@ -24,22 +22,6 @@ test("2B. Should access DRep Directory page", async ({ page }) => {
       page.getByRole("navigation").getByText("DRep Directory")
     ).toBeVisible();
   }
-});
-
-test("2I. Should check validity of DRep Id", async ({ page }) => {
-  const dRepDirectory = new DRepDirectoryPage(page);
-  await dRepDirectory.goto();
-
-  await dRepDirectory.searchInput.fill(dRep01Wallet.dRepId);
-  await expect(dRepDirectory.getDRepCard(dRep01Wallet.dRepId)).toHaveText(
-    dRep01Wallet.dRepId
-  );
-
-  const wallet = await ShelleyWallet.generate();
-  const invalidDRepId = extractDRepFromWallet(wallet);
-
-  await dRepDirectory.searchInput.fill(invalidDRepId);
-  await expect(dRepDirectory.getDRepCard(invalidDRepId)).not.toBeVisible();
 });
 
 test("2D. Should show delegation options in connected state", async ({
